test(eslint-config): refactor integration tests to include more rules
So, the idea is to fix:

    https://github.com/liferay/liferay-frontend-projects/issues/515

And in order to do so, I have to make the integration tests a little bit
more representative of what happens in liferay-portal. So, that means
turning on the `lines-around-comment` rule, among others.

Set-up is a bit more complicated, but also more true to reality, so I
guess that's a win? If you ever want to add another standard ESLint rule
that's not already in the base config, you can call
`addRule('the-rule')`.

Test plan: Run tests.