fix(eslint-config): fix edge case with comment handling
Comment handling here is a bit edge-casey, so I rewrote it to be a bit
more comprehensive, thus solving the failing regression test that I
added in the parent comment.

The old code assumed only one comment could/would be in the "leading"
position, but the regression test shows how two comments break it.

So we actually iterate over all the preceding comments now identifying
which ones are considered to be "leading":

    // This is a leading comment.
    // This one is too.
    one();

    two(); // Trailing comment.
    // Another leading comment.
    three();

ie. we have to take special care to ensure that the "Trailing comment"
there doesn't get considered to belong with the "Another leading
comment" that comes right after it.

I am reasonably confident that this works pretty generally now, because
as I was writing it, I saw how my initial attempt failed to handle other
cases that are already represented in the test suite (like the one
above), but with the final version everything remains green.

Test plan: Run tests. Copy this to liferay-portal and see if it fixes
Bryce's bug.

Closes: https://github.com/liferay/liferay-frontend-projects/issues/515