refactor(eslint-config): remove redundant `addRule()` call
Meant to include this in a prior commit but somehow it got dropped on
the floor. As you can see from the comment, in an intermediate version
of this PR:

    https://github.com/liferay/liferay-frontend-projects/pull/517

I was thinking that I would add some rules to the base set selectively,
to make all this resemble the liferay-portal set-up a bit more
realistically. I didn't think that I could add _all_ the rules without
stuff breaking.

In the end though, I could: nothing breaks, so we may as well turn them
all on. I just forgot to clean up this straggling comment and call.

Test plan: Run unit tests; they are still green.