fix(eslint-config): handle one more edge case
Based on testing in liferay-portal, there is a subtle difference between
the environment there and the one in here, as seen when testing:

    https://github.com/liferay/liferay-frontend-projects/pull/517

Basically, the tests were green here, but when I copied the code over
there, I still saw the unwanted lint errors about:

    imports must be grouped (unexpected blank line before: "frontend-js-web")

Something to due with rule order, or the defaults that come in via
`npm-scripts`.

Whatever the cause, the problem was that the preceding code didn't cope
well with whitespace when it came to identifying "leading" comments. A
"leading" comment is a comment that goes before a thing and applies to
that thing.

    // This is a leading comment.
    thing();

    // This one is too, despite the blank line that follows:

    thing();

    // Humans consider all of these lines to be "a comment", but
    // they are technically multiple instances of "line" comments,
    // and ESLint treats them as such.

    thing();

    // This is not a leading comment.

    // But this one is.

    thing();

    thing(); // This is a trailing comment, not a leading one.
    // This one is a leading comment.

    thing();

Anyway, some small tweaks to the code to handle these cases. The
nutshell version is that something is considered a leading comment if it
is last comment before a thing, or there are no blank lines between it
and the next comment, and in both cases we make sure it isn't a
"trailing" comment (ie. following a thing on the same line as that
thing).

Now, there are still more edge cases that we _could_ handle here, like
these two shouldn't be considered part of the same "comment":

    /* one */
    // two

    thing();

But I am punting on that until we actually see such an odd construction
out in the wild.

Test plan: Copy file into `liferay-portal` and test it out on
`frontend-js-tooltip-support-web`. I tweaked the test a bit so show that
we can autoformat to the expected shape. Didn't encode it in the test
suite but I manually tested a bunch of the combinations shown above to
see that it behaved reasonably.