feat(js-toolkit): implement `npm run start eject`
This is a new sub-target that can be run once to detach the webpack configuration associated to
`npm run start` to further tweak it.

The idea behind this feature is that users willing to make `npm run start` more powerful can eject
the configuration so that the JS Toolkit does not regenerate prior to each run of `npm run start`.

This way, people wanting to mock the Liferay object or tweaking webpack configuration don't
need to send a PR and wait for another JS Toolkit release and/or pass through all the review
process.

It will also stop preventing users from doing things that cannot be generalized for everybody and,
thus don't fit in the JS Toolkit at all.

Fixes #235