fix(npm-scripts): mock CSS files via resolver (instead of transformer)
Because we need to import some CSS files living in the global
`node_modules` and because Jest configures `/node_modules/` as one of its
default `transformIgnorePatterns`, we are changing from letting Jest
resolve CSS files to their real paths and then transforming them to the
empty string to resolve all CSS files to an empty CSS file.

This way we achieve the same result without the need to mess around with
the `transformIgnorePatterns`.

See https://github.com/liferay/liferay-frontend-projects/issues/534 for
more details on the rationale of this decission.

Fixes #534