feat(target-platforms): add a node script to create target platforms
This script receives the path to a liferay-portal repository and the name
of the target platform and creates a new target platform project after
scanning liferay-portal sources and node_modules folder.

It basically reads the modules/npmscripts.config.js file to generate the
target platform imports and then relies on require.resolve invocations
inside the liferay-portal sources to find the version numbers of each
dependency listed as import.

As of now, this works for master and 7.4.0-GA1. It's possible that we need
to amend it in the future to support older and newer versions.