chore(target-platforms): get target platforms out of yarn workspaces
It doesn't make sense to include target platforms in a yarn workspace
because, by definition, most of the dependencies are not meant to be
hoisted (not even fetched, because they are not needed for anything during
build time).

Thus, if we add them to the yarn workspaces, the number of dependencies
will grow very fast when new releases of Liferay are published, making the
repository hard to manage for no benefit.

Note that we still add the target-platforms and create-platform projects to
the yarn workspace to benefit from liferay-workspace-scripts.