feat(js-toolkit): implement autopreset support
This feature makes the `liferay-npm-bundler` look for direct project
dependencies that are presets and contain a direct dependency to
`liferay-npm-bundler` (i.e: presets that pull the bundler with them so that
the project doesn't need to add it to its list of dependencies).

When one is found, it is automatically applied if no explicit preset is
defined in the project.

If more than one is found and the .npmbundlerrc file does not specify which
one to load, the build fails.