feat(liferay-cli): add command to create a new project
This creates a new project directory with the given name and containing
the .gitignore, README.md and package.json files for now.

It is the initial stage of the creator. Now we need to add sample source
code, localization, and configuration examples.

The structure of the code is based on the current yeoman generator, which
is composed by facets and targets.

A target is a type of project than can be generated, like: Liferay Platform
project, Angular, React, ...

A facet is a common feature of a project that can be applied to more than
one target.

This commits implements the facet-project facet that creates the folder and
core project files and the target-liferay target that configures the target
platform on top of what facet-project has generated.

Given that yeoman uses `inquirer` under the hood to show prompts, I've
added a dependency to use it because, implementing it from scratch was
probably a bit overkill.

I'm pasting a sample of the execution and contents of generated project
below:

someone@somewhere:~ $ liferay new my-app
👋 Welcome to Liferay Project Generator v1.0.0
🛈 Gathering info to create new project 'my-app'...
? What is the human readable description of your project? My App
? What will be your target platform? portal-7.4-ga1
🛈 Generating new project 'my-app'...
✅ Successfully generated new project 'my-app'

  You can now run the following commands to build your project:

      $ cd my-app ↩
      $ yarn build ↩

  This will create a my-app.jar file in your my-app/dist
  folder that you can deploy to your local Liferay server.

someone@somewhere:~ $ find my-app
my-app
my-app/.gitignore
my-app/package.json
my-app/README.md
my-app/assets
my-app/assets/.placeholder

someone@somewhere:~ $ cat my-app/package.json
{
  "name": "my-app",
  "version": "1.0.0",
  "description": "My App",
  "dependencies": {
    "portal-7.4-ga1": "^1.0.0"
  },
  "scripts": {
    "build": "liferay build"
  }
}

someone@somewhere:~ $ _