feat(liferay-cli): use npm in place of yarn (as it is more standard)
Also remind user about `npm install` step before building.

I was tempted to add an automatic step for this, but then though that it's
better to leave it to the user because:

1. We don't know if the user wants to use npm or yarn
2. We don't know if the project is being generated inside a yarn workspace
3. The user may want to `yarn link` dependencies or may have them
   yarn-linked in the root of the workspace and running `yarn` again may
   lead to an overwrite of those.
4. Yeoman generators usually run `npm install` for you and that leads to a
   very cluttered display of actions, making it difficult to know what has
   been done.
5. Sometimes `npm install` fails and, again, that leads to a cluttered
   display and forces the tool to warn the user about that circumstance and
   ask him to run it again.

So, unless someone sees a huge benefit in doing `npm install` automatically
I would say we leave it like this.