feat(liferay-cli): generate sample code for target-liferay
Generate configuration declaration for portlet instance and System Settings
levels.

Add CSS styles and L10N keys.

This sample generation used to be modular in the old generator, but I'm
leaning towards making it more hard-coded in this new tool because it is
not modified so often and we don't see big gains because of modularity.

Some parts of the code will need to be refactored because they will need to
be used in other targets (React, Angular, Vue, ... portlets) but we will do
that when it is really needed.