refactor(changelog-generator): no need to use `fs` async API
Given that `fs` has sync versions of every file operation, we don't need to
promisify anything and the `await` on it.

Also, in tests I made for liferay-npm-bundler long ago, I found out that
parallelizing reads when transforming JS files did not lead to any increase
in performance (as counterintuitive as it may seem).

I suspect that this is because jumping accross the whole disk to make
random reads in different files leads to less throughput than reading them
serially one by one so, unless you have to do anything else while waiting
for I/O, there's no point in parallelizing anything.

In the case where the OS has the files cached in memory, you don't need to
jump accross the disk, but still reading things serially may be more
efficient because you reduce the overhead and invocations of reads are
done instantly.