chore: put target-platforms in its own workspace
Since every platform has its own set of dependencies and we are moving the
majority of `create-platform` assets to a shared package, there's no point
in adding target platforms to the global yarn workspace.

We will still use a yarn workspace for target platforms only, to share
development dependencies that may make sense and to be able to leverage
yarn for maintenance tasks.