feat(portal-base): add deploy command
We first refactor target platforms to extract common code to a shared
package so that releasing new platforms is easier in the future, per these
comment and issue:

https://github.com/liferay/liferay-frontend-projects/pull/584#discussion_r651344473
https://github.com/liferay/liferay-frontend-projects/issues/591

Note that we leave liferay-npm-bundler as a dependency of the target
platform and declare it as a peerDependency in the common package. This is
because liferay-npm-bundler is the autopreset marker so we cannot move it
to the portal-base as liferay-npm-bundler would fail to recognize
portal-7.4-ga1 as an autopreset.

We may change the bundler and portal-base in the future if the above poses
any problem.