feat(portal-base): add per-project local configuration helper
The configuration.js module deals with local per-project configuration
contained in .liferay.json. This information was previously stored in
.npmbuildrc, but since liferay-cli is new, we use this new naming, which
seems more aligned to what it does.

For now we will simply store the path to the local Liferay installation,
but we may use it in the future if we need to store more local settings
that must not be committed to the CVS.