refactor(target-platforms): specify liferay-npm-bundler version
Move specification of liferay-npm-bundler version to use to portal-base
(from target platforms) so that we don't need to specify it in different
places when releasing to npmjs.com.

Leaving liferay-npm-bundler version number as * in portal-7.4-ga1 does the
trick because it will always depend on portal-base, which is the project
driving what version to use.