feat(target-platforms): create portal-master target platform
This is intended to be used for testing purposes only, not by the general
public. It should never be released to npmjs.com (unless as a snapshot
version) and it may be obsolete at any time (depending on liferay-portal
changes) so it is wise to update it with the `update-platforms` command
before using it.