test(target-platforms): account for machine-specific tokens in QA tests
For now, we only need to substitute WORKSPACE_DIR in `.css.map` files to
make QA tests machine agnostic, but in the future we may need to add more
logic to the `readExpected` method, or even refactor the whole `diff`
logic to make it smarter.

But let's go step by step.