feat(dev-server): Maps modules in liferay-portal to paths in the server
We use a very naive approach for now and just map the name of the module
from its `package.json` to the `build/node/packageRunBuild/resources` folder
in the WD.

At runtime, we know requests for those resources should look more or less
like `/o/${moduleName}/${path}`, and those can be retrieved (when built)
from `build/node/packageRunBuild/resources/${path}`.

This does not account for `.scss` files which are compiled to `.css` and
placed in `.sass-cache`, so we'll need to further refine this to account
for it.

Additionally, globbing `liferay-portal` is a very expensive operation. We
could try to defer it somehow, but I'm not sure we can have a functioning
dev-server without it. Thought about doing a kind of reverse-lookup as
requests started coming in and then slowly fill the map on demand, but not
sure how much of a win that could be if at all, so paying the price upfront
for now.