feat(dev-server): Sets up a basic proxy to return mapped files from disk when available
Opens a basic server at 127.0.0.1:3000 that acts as a proxy to the local
server when the requests are not part of our mapped module files.

For now, we only handle JS for:
- /o/moduleName/fileName
- /o/js/resolved-module/moduleName/fileName

AUI modules can't be found in their expected `build/node/packageRunBuild`
location since they're only added via Gradle. We'd need to map them and
other similar files in a different way.