chore: remove js-toolkit/qa projects from yarn workspace
This is necessary since we cannot use local versions of the build tools
obtained from the links in the yarn workspace because that would make
resolution of artifacts unstable.

Instead, we will rely on local npm registries (like Verdaccio) to test QA
projects in a controlled version environment. The setup for that is yet to
come, but on its way.