chore(target-platform): let portal-master be published
We need to be able to publish this platform to local npm registries to be
able to test it. Also, in case we published it to npmjs.com nothing
would happen (but still, it's almost impossible to publish it because you
need special credentials to be able to publish for the first time).