refactor(liferay-cli): make facets independent of each other
Instead of testing for file existence, it's better to test for options.

This couples facets between them too, but instead of relying in the order
they are invoked, it relies on them being present on the generation, which
is a much weaker and more reasonable assumption.

So, for example, if factetReact.render() depends on facet-localization's
options to know whether it should generate L10N labels, it is better to
check options.addLocalizationSupport because it will always exist as long
as facetLocalization.prompt() has been invoked, which is a very reasonable
assumption and, in any case, better than assuming that
facetLocalization.render() is always called before facetReact.render().