chore(target-platforms): reorganize dependencies
Move all JS Toolkit to portal-base and inherit it from portal-adapt-base.

Note that portal-base could be split in two because it contains
dependencies for the JS Toolkit but for portal projects builds too,
something that may not be needed in customer projects or adaptations.

However, since those deps are babel and sass and we will probably want to
invoke them from any type of project, I'm leaving it as one single project
for now. Time will tell...