feat(js-toolkit-core): don't warn user when loading old bundler projects
I added this check so that bundler 3 users were warned about .npmbundlerrc
files. This made sense because we were going to make people upgrade soon,
but since we delayed bundler 3 it is no longer the case.

Now, it will not warn unless `Project.WARN_ON_NPMBUNDLERRC` is set to
`true` before using it.