refactor(liferay-cli): don't import scripts, but lazily require them
This is to avoid a nasty error when loading './new' inside old adapted
projects (this happens when you are upgrading them) that makes the
application fail without showing any meaningful error.

By lazy requiring the scripts, we only load them when they are going to be
executed, preventing the aforementioned error.