feat(npm-scripts): support ESM exports
For now we create bridges from AMD modules to ESM ones so that they can be
used with native `import from` syntax.

In the future we may need to rewrite this, if we can get rid of the AMD
loader or see that's better doing it the other way (i.e: make ESM modules
available as AMD modules).