feat(portal-base): move manifest.json generation to webpack phase
We emit the manifest.json file from a webpack plugin so that it can be
fetched from the live development server too. Otherwise, the live
development process is very cumbersome because there's no way to register
the bundled app in the portal without deploying it to a real server at
least once.