test(js-toolkit): add QA support for remote apps
Given that remote apps cannot be deployed (only started), we check their
output with a expected project directory that contains the expected output
from the build process.

For now we only compare file by file and dir by dir, but in the future we
may add tokens and smarter logic to perform the diff.