feat(npm-scripts): let exports be bridged
With this we allow a new syntax for exports. Instead of being a single
string with a package name, we can now use a richer syntax:

```javascript
{
  package: string,
  bridge: boolean,
}
```

Where bridge can be true to force npm-scripts to create a temporary bridge
to workaround incorrect webpack export under some circumstances, described
below:

(taken from https://github.com/liferay-frontend/liferay-portal/pull/1955#issuecomment-1057983985)

OK, so, the reason why this was failing is that `@clayui/provider`
[is using some way to export its stuff](https://github.com/liferay/clay/blob/c62913222b959af3ce13f0e00cd10a8406ab4a33/packages/clay-provider/src/index.ts#L6) that when processed by typescript is translated to [something obscure](https://unpkg.com/@clayui/provider@3.40.0/lib/index.js)
that is probably fooling webpack.

This code is not statically analyzable:

```javascript
Object.keys(_Provider).forEach(function (key) {
  if (key === "default" || key === "__esModule") return;
  if (key in exports && exports[key] === _Provider[key]) return;
  Object.defineProperty(exports, key, {
    enumerable: true,
    get: function get() {
      return _Provider[key];
    }
  });
});
```
Even though it comes from
[one that was statically analizable](https://github.com/liferay/clay/blob/c62913222b959af3ce13f0e00cd10a8406ab4a33/packages/clay-provider/src/index.ts#L6).

Because of that, webpack seems to give up when trying to determine what
to export with ESM syntax and the produced .js file ends up exporting
nothing, thus making everything fail at runtime.

One possible way of fixing this would be generating ESM code while
constructing `@clayui/provider` (in addition to the CommonJS one) and
using that from webpack.

However, given that we can expect to find this problem in more places, I
have created some code in `npm-scripts` to workaround this.

I will release a new snapshot of `npm-scripts` and push some changes to
this PR to use it and see what happens. In any case, I've tried the fix
locally and the icons are not lost any more.