feat(npm-scripts): support exports adaptation for CommonJS packages
As the current npm ecosystem is built, there are some npm packages that
export things with ES syntax (export) while others export things using
CommonJS syntax (module.exports).

Everything should slowly swith to ES syntax, but, for instance, react is
still using module.exports for some of its packages.

The problem is that webpack (or any equivalent tool in this universe)
cannot always infer what is exported through module.exports. This is
because ES's export keyword is a declaration (can be analyzed statically at
build time) but module.exports needs to be analyzed at runtime to see
what's.

This commit adds some code to infer what is exported through module.exports
by requiring the package from npm-scripts and inspecting what is returned
(that's what symbols: 'auto' means in any npmscripts.config.js exports
configuration).

This mostly works, but still some packages intended to be run exclusively
in the browser may fail to be required in node. For those packages we would
have to manually describe what is exported providing symbols with an array,
instead of the word 'auto'.