refactor(npm-scripts): change getBndWebContextPath signature
I want to slowly convert all functions to stop using global variables
(like what getMergedConfig() returns) or state (like assuming the CWD is
the project dir).

This is because we did the same thing in JS Toolkit and it proved to be a
very bad decision because it makes testing harder, makes the functions not
reentrant, and create a bunch of issues.

So, what I had in JS Toolkit was an object that could be passed through
functions and contained the state and description of the project, but we
may as well use the projectDir and CONFIG objects for that in npm-scripts.
I think that will be enough, most probably, and won't require a huge
refactor. We'll see.