feat(amd-loader): support useESM flag
This flag will be set by npm-scripts whenever any AMD module needs to be
loaded with <script type=module>. This was not the previous behaviour, but
we need it to interoperate ESM and AMD modules via bridges.

Note that the URL builder has to differentiate now between plain JS and
ES modules when creating combo URLs. This is because we cannot mix ES and
JS modules in the same combo request as they would be loaded by just one
<script> tag which would need to be either <script> or
<script type="module"> (but not both) as soon as the wave function
collapses.