chore(amd-loader): fix yarn proxyPortal
To be able to use webpack-dev-server we need to avoid hoisting from yarn
and make sure we use an old version (since we haven't evolved this project
for a long time).

I'm also changing the default path to loader.js in Liferay DXP, since it
change some time ago from frontend-js-web to
frontend-js-loader-modules-extender.