feat(amd-loader): update resolve path when a redirect is received
Due to https://issues.liferay.com/browse/LPS-169194, we have make the
/o/js_resolve_modules endpoint cacheable forever, so that CDNs and browsers
can take advantage of it.

With the new scheme, the /o/js_resolve_modules URL gets a UUID appended
that is changed every time the server's NPMRegistry state changes.

Since the loader is cached in the JS runtime engine, we cannot update the
resolve path from the server when it changes, because we cannot push code.
However, we may send a redirect to the AMD loader to tell it where's the
new endpoint.

Note that the old AMD loader will still work with the new server (even in
absence of this optimization).This change is only an improvement for the
case when the resolve path changes and the loader still wants to make
more calls to the resolve modules endpoint: by updating the cached resolve
path, we make sure that we need to follow the redirect from the server just
once, saving a lot of duplicated requests due to redirects.