feat: support internal paths in exports
Note that we only support .css and .js exports for now. In the future we
may want to support .json, .svg, or other static resources we see fit.

This is part of [https://issues.liferay.com/browse/LPS-172747](LPS-172747).

----
⚠ WARNING !!!

This implementation is missing a corner use case: an AMD module that wants
to use an internal path export ending in .js. We won't support that case
unless it is absolutely necessary, since it can be fixed migrating the
offending project from AMD to ESM.
----