feat: allow new exports syntax for internal paths
Now we will be able to do:

exports: [
  {
    name: 'data-set',
    path: './src/main/resources/META-INF/resources/data-set/index.ts'
  },
]

And that will create an export under __liferay__, named 'data-set.js',
containing the webpack bundle for the configured entry point.

This will make it possible to export more modules in addition to the "main"
one, which will be useful for our publics API OSGi module.

Note that this commit moves the logic to infer the file paths to the
configuration normalizer and assigns it to the field "name".