refactor: extract common build process code
Since we will be doing almost the same for all CX types, I want to reuse
as much as possible and make all targets have the same structure, so that
we can generalize them in the near future and make contribution of new CX
types even easier.

For now I have detected that build processes share:

1. A check configuration step
2. Step for copying static assets
3. The specific build (which most of the times will be webpack)
4. Making the ZIP file

And steps 2 and 4 are the same for all CX types, for example, so they can
be extracted to a common framework.