feat: add global `Liferay` object API
This is a PoC that exports the current Liferay.ThemeDisplay API as an
export named `liferay` inside `@liferay/js-api`.

In general, for every stable release of `@liferay/js-api`, the API here
should reflect the real structure of one or more successive versions of
DXP.

We may need to add some npm prepublish checks to avoid releasing incorrect
versions of these files. For that we can leverage the comment that is added
at the top of the files showing the portal version from where they came.