feat(npm-scripts): support SVG exports
They work similar to CSS but inline the content directly instead of
retrieving it from the server. This is because SVG files don't undergo any
process in the server, as CSS files do, so there's no need to make two
trips to the server when we can make just one.